Rename level_dat.rs to settings.rs
